% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebppackages-package.R
\docType{package}
\name{ebppackages-package}
\alias{ebppackages}
\alias{ebppackages-package}
\title{ebppackages: Best Practices for Using eBird Data}
\description{
Metapackage for the book Best Practices for using
    eBird Data and the associated paper. Install this package to install
    all the dependencies for the book.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mstrimas/ebird-best-practices}
}

}
\author{
\strong{Maintainer}: Matthew Strimas-Mackey \email{mes335@cornell.edu} (0000-0001-8929-7776)

}
\keyword{internal}
